+++ flutter_local_notifications_platform_interface/CHANGELOG.md
@@ -1,3 +1,7 @@
+## vNext
+
+* Added web platform support to the platform interface
+
 ## [11.0.0-dev.1]
 
 * **Breaking change** bumped minimum Flutter SDK requirement to 3.38.1 and Dart SDK requirement to 3.10.0